Move HiveParquetColumnIOConverter#constructField to trino-parquet
The logic here does not have dependency on hive and it is needed
here to allow ParquetWriter to create ParquetReader for the
verification of the written file.
