--- plugin/trino-hive/src/main/java/io/trino/plugin/hive/parquet/HiveParquetColumnIOConverter.java
@@ -1,92 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.trino.plugin.hive.parquet;
-
-import com.google.common.collect.ImmutableList;
-import io.trino.parquet.Field;
-import io.trino.parquet.GroupField;
-import io.trino.parquet.PrimitiveField;
-import io.trino.spi.type.ArrayType;
-import io.trino.spi.type.MapType;
-import io.trino.spi.type.RowType;
-import io.trino.spi.type.Type;
-import org.apache.parquet.io.ColumnIO;
-import org.apache.parquet.io.GroupColumnIO;
-import org.apache.parquet.io.PrimitiveColumnIO;
-
-import java.util.List;
-import java.util.Locale;
-import java.util.Optional;
-
-import static io.trino.parquet.ParquetTypeUtils.getArrayElementColumn;
-import static io.trino.parquet.ParquetTypeUtils.getMapKeyValueColumn;
-import static io.trino.parquet.ParquetTypeUtils.lookupColumnByName;
-import static org.apache.parquet.io.ColumnIOUtil.columnDefinitionLevel;
-import static org.apache.parquet.io.ColumnIOUtil.columnRepetitionLevel;
-import static org.apache.parquet.schema.Type.Repetition.OPTIONAL;
-
-public final class HiveParquetColumnIOConverter
-{
-    private HiveParquetColumnIOConverter() {}
-
-    public static Optional<Field> constructField(Type type, ColumnIO columnIO)
-    {
-        if (columnIO == null) {
-            return Optional.empty();
-        }
-        boolean required = columnIO.getType().getRepetition() != OPTIONAL;
-        int repetitionLevel = columnRepetitionLevel(columnIO);
-        int definitionLevel = columnDefinitionLevel(columnIO);
-        if (type instanceof RowType) {
-            RowType rowType = (RowType) type;
-            GroupColumnIO groupColumnIO = (GroupColumnIO) columnIO;
-            ImmutableList.Builder<Optional<Field>> fieldsBuilder = ImmutableList.builder();
-            List<RowType.Field> fields = rowType.getFields();
-            boolean structHasParameters = false;
-            for (int i = 0; i < fields.size(); i++) {
-                RowType.Field rowField = fields.get(i);
-                String name = rowField.getName().orElseThrow().toLowerCase(Locale.ENGLISH);
-                Optional<Field> field = constructField(rowField.getType(), lookupColumnByName(groupColumnIO, name));
-                structHasParameters |= field.isPresent();
-                fieldsBuilder.add(field);
-            }
-            if (structHasParameters) {
-                return Optional.of(new GroupField(type, repetitionLevel, definitionLevel, required, fieldsBuilder.build()));
-            }
-            return Optional.empty();
-        }
-        if (type instanceof MapType) {
-            MapType mapType = (MapType) type;
-            GroupColumnIO groupColumnIO = (GroupColumnIO) columnIO;
-            GroupColumnIO keyValueColumnIO = getMapKeyValueColumn(groupColumnIO);
-            if (keyValueColumnIO.getChildrenCount() != 2) {
-                return Optional.empty();
-            }
-            Optional<Field> keyField = constructField(mapType.getKeyType(), keyValueColumnIO.getChild(0));
-            Optional<Field> valueField = constructField(mapType.getValueType(), keyValueColumnIO.getChild(1));
-            return Optional.of(new GroupField(type, repetitionLevel, definitionLevel, required, ImmutableList.of(keyField, valueField)));
-        }
-        if (type instanceof ArrayType) {
-            ArrayType arrayType = (ArrayType) type;
-            GroupColumnIO groupColumnIO = (GroupColumnIO) columnIO;
-            if (groupColumnIO.getChildrenCount() != 1) {
-                return Optional.empty();
-            }
-            Optional<Field> field = constructField(arrayType.getElementType(), getArrayElementColumn(groupColumnIO.getChild(0)));
-            return Optional.of(new GroupField(type, repetitionLevel, definitionLevel, required, ImmutableList.of(field)));
-        }
-        PrimitiveColumnIO primitiveColumnIO = (PrimitiveColumnIO) columnIO;
-        return Optional.of(new PrimitiveField(type, required, primitiveColumnIO.getColumnDescriptor(), primitiveColumnIO.getId()));
-    }
-}
